+++ build_runner/lib/src/generate/watch_impl.dart
@@ -9,6 +9,7 @@ import 'package:build_runner/src/watcher/asset_change.dart';
 import 'package:build_runner/src/watcher/graph_watcher.dart';
 import 'package:build_runner/src/watcher/node_watcher.dart';
 import 'package:crypto/crypto.dart';
+import 'package:glob/glob.dart';
 import 'package:logging/logging.dart';
 import 'package:stream_transform/stream_transform.dart';
 import 'package:watcher/watcher.dart';
@@ -23,6 +24,7 @@ import '../environment/overridable_environment.dart';
 import '../package_graph/apply_builders.dart';
 import '../package_graph/build_config_overrides.dart';
 import '../package_graph/package_graph.dart';
+import '../package_graph/target_graph.dart';
 import '../server/server.dart';
 import '../util/constants.dart';
 import 'build_definition.dart';
@@ -53,6 +55,8 @@ Future<ServeHandler> watch(
   Map<String, BuildConfig> overrideBuildConfig,
 }) async {
   packageGraph ??= new PackageGraph.forThisPackage();
+  final targetGraph = await TargetGraph.forPackageGraph(packageGraph,
+      overrideBuildConfig: overrideBuildConfig);
   var environment = new OverrideableEnvironment(
       new IOEnvironment(packageGraph, assumeTty),
       reader: reader,
@@ -63,6 +67,7 @@ Future<ServeHandler> watch(
       deleteFilesByDefault: deleteFilesByDefault,
       failOnSevere: failOnSevere,
       packageGraph: packageGraph,
+      rootPackageConfig: targetGraph.rootPackageConfig,
       logLevel: logLevel,
       debounceDelay: debounceDelay,
       skipBuildScriptCheck: skipBuildScriptCheck,
@@ -70,8 +75,7 @@ Future<ServeHandler> watch(
   var terminator = new Terminator(terminateEventStream);
 
   overrideBuildConfig ??= await findBuildConfigOverrides(options.packageGraph);
-  final buildActions = await createBuildActions(options.packageGraph, builders,
-      overrideBuildConfig: overrideBuildConfig);
+  final buildActions = await createBuildActions(targetGraph, builders);
 
   var watch =
       runWatch(environment, options, buildActions, terminator.shouldTerminate);
@@ -95,13 +99,15 @@ Future<ServeHandler> watch(
 /// run.
 WatchImpl runWatch(BuildEnvironment environment, BuildOptions options,
         List<BuildAction> buildActions, Future until) =>
-    new WatchImpl(environment, options, buildActions, until);
+    new WatchImpl(environment, options, buildActions, until,
+        options.rootPackageFilesWhitelist.map((g) => new Glob(g)));
 
 typedef Future<BuildResult> _BuildAction(List<List<AssetChange>> changes);
 
 class WatchImpl implements BuildState {
   AssetGraph _assetGraph;
   BuildDefinition _buildDefinition;
+  final Iterable<Glob> _rootPackageFilesWhitelist;
 
   /// Delay to wait for more file watcher events.
   final Duration _debounceDelay;
@@ -124,8 +130,12 @@ class WatchImpl implements BuildState {
   final _readerCompleter = new Completer<DigestAssetReader>();
   Future<DigestAssetReader> get reader => _readerCompleter.future;
 
-  WatchImpl(BuildEnvironment environment, BuildOptions options,
-      List<BuildAction> buildActions, Future until)
+  WatchImpl(
+      BuildEnvironment environment,
+      BuildOptions options,
+      List<BuildAction> buildActions,
+      Future until,
+      this._rootPackageFilesWhitelist)
       : _directoryWatcherFactory = environment.directoryWatcherFactory,
         _debounceDelay = options.debounceDelay,
         packageGraph = options.packageGraph {
@@ -283,7 +293,7 @@ class WatchImpl implements BuildState {
 
   bool _isWhitelistedPath(AssetId id) {
     if (packageGraph[id.package].isRoot) {
-      return rootPackageGlobsWhitelist.any((glob) => glob.matches(id.path));
+      return _rootPackageFilesWhitelist.any((glob) => glob.matches(id.path));
     } else {
       return id.path.startsWith('lib/');
     }