+++ build_runner/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.7.3+1-dev
+## 0.7.4
 
+- Allows using files in any build targets in the root package as sources if they
+  fall outside the hardcoded whitelist.
 - Changes to the root `.packages` file during watch mode will now cause the
   build script to exit and prompt the user to restart the build.